fix: move remote hooks
